Renan

+++ src/main/java/com/betha/statustce/statustce/model/GrupoAssunto.java
@@ -11,20 +11,20 @@ public class GrupoAssunto {
 
     @Id
     @Column(name="id")
-    private Integer id;
+    private Long id;
     @Column(name="descricao")
     private String descricao;
 
-    public GrupoAssunto(Integer id, String descricao) {
+    public GrupoAssunto(Long id, String descricao) {
         this.id = id;
         this.descricao = descricao;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
